{{ cookiecutter.repo_name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author_name }}"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
mlflow = "^2.0.1"
bentoml = "^1.0.12"
prefect = "^2.7.1"
scikit-learn = "^1.2.0"
hydra-core = "^1.3.0"
altair = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79